FE changes for table extraction
